--- x-pack/solutions/observability/plugins/streams/server/routes/streams/delete.ts
@@ -1,66 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import { badRequest, internal, notFound } from '@hapi/boom';
-import { z } from '@kbn/zod';
-import {
-  DefinitionNotFound,
-  ForkConditionMissing,
-  IndexTemplateNotFound,
-  SecurityException,
-} from '../../lib/streams/errors';
-import { MalformedStreamId } from '../../lib/streams/errors/malformed_stream_id';
-import { createServerRoute } from '../create_server_route';
-
-export const deleteStreamRoute = createServerRoute({
-  endpoint: 'DELETE /api/streams/{id}',
-  options: {
-    access: 'internal',
-  },
-  security: {
-    authz: {
-      enabled: false,
-      reason:
-        'This API delegates security to the currently logged in user and their Elasticsearch permissions.',
-    },
-  },
-  params: z.object({
-    path: z.object({
-      id: z.string(),
-    }),
-  }),
-  handler: async ({
-    params,
-    logger,
-    request,
-    getScopedClients,
-  }): Promise<{ acknowledged: true }> => {
-    try {
-      const { streamsClient } = await getScopedClients({
-        request,
-      });
-
-      await streamsClient.deleteStream(params.path.id);
-
-      return { acknowledged: true };
-    } catch (e) {
-      if (e instanceof IndexTemplateNotFound || e instanceof DefinitionNotFound) {
-        throw notFound(e);
-      }
-
-      if (
-        e instanceof SecurityException ||
-        e instanceof ForkConditionMissing ||
-        e instanceof MalformedStreamId
-      ) {
-        throw badRequest(e);
-      }
-
-      throw internal(e);
-    }
-  },
-});
