--- x-pack/solutions/observability/plugins/streams/server/routes/streams/edit.ts
@@ -1,62 +1,0 @@
-/*
- * Copyright Elasticsearch B.V. and/or licensed to Elasticsearch B.V. under one
- * or more contributor license agreements. Licensed under the Elastic License
- * 2.0; you may not use this file except in compliance with the Elastic License
- * 2.0.
- */
-
-import { badRequest, internal, notFound } from '@hapi/boom';
-import { streamConfigDefinitionSchema } from '@kbn/streams-schema';
-import { z } from '@kbn/zod';
-import {
-  DefinitionNotFound,
-  ForkConditionMissing,
-  IndexTemplateNotFound,
-  RootStreamImmutabilityException,
-  SecurityException,
-} from '../../lib/streams/errors';
-import { MalformedStreamId } from '../../lib/streams/errors/malformed_stream_id';
-import { createServerRoute } from '../create_server_route';
-
-export const editStreamRoute = createServerRoute({
-  endpoint: 'PUT /api/streams/{id}',
-  options: {
-    access: 'internal',
-  },
-  security: {
-    authz: {
-      enabled: false,
-      reason:
-        'This API delegates security to the currently logged in user and their Elasticsearch permissions.',
-    },
-  },
-  params: z.object({
-    path: z.object({
-      id: z.string(),
-    }),
-    body: streamConfigDefinitionSchema,
-  }),
-  handler: async ({ params, request, getScopedClients }) => {
-    try {
-      const { streamsClient } = await getScopedClients({ request });
-      const streamDefinition = { stream: params.body, name: params.path.id };
-
-      return await streamsClient.upsertStream({ definition: streamDefinition });
-    } catch (e) {
-      if (e instanceof IndexTemplateNotFound || e instanceof DefinitionNotFound) {
-        throw notFound(e);
-      }
-
-      if (
-        e instanceof SecurityException ||
-        e instanceof ForkConditionMissing ||
-        e instanceof MalformedStreamId ||
-        e instanceof RootStreamImmutabilityException
-      ) {
-        throw badRequest(e);
-      }
-
-      throw internal(e);
-    }
-  },
-});
